requirements.txt
arabic-reshaper==2.1.3
asgiref==3.4.1
asn1crypto==1.5.1
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.2
cssselect2==0.6.0
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0.3
django-admin-interface==0.19.0
django-allauth==0.51.0
django-colorfield==0.6.3
django-constance==2.9.0
django-cors-headers==3.11.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-picklefield==3.1
django-resized==1.0.2
django-user-visit==0.5.1
djangorestframework==3.13.1
future==0.18.2
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
lxml==4.9.1
more-itertools==8.13.0
oauth2client==4.1.3
oauthlib==3.2.0
oscrypto==1.3.0
packaging==21.3
Pillow==9.2.0
pluggy==0.13.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyHanko==0.13.2
pyhanko-certvalidator==0.19.5
PyJWT==2.4.0
pyparsing==3.0.9
PyPDF3==1.0.6
pytest==5.4.1
pytest-django==3.9.0
python-barcode==0.13.1
python-bidi==0.4.2
python-decouple==3.0
python3-openid==3.2.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.11
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.20.0
rsa==4.9
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
static3==0.7.0
svglib==1.3.0
tinycss2==1.1.1
toml==0.10.2
tqdm==4.64.0
tzdata==2021.5
tzlocal==4.2
ua-parser==0.15.0
uritools==4.0.0
urllib3==1.26.9
user-agents==2.2.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
xhtml2pdf==0.2.8
